src/requirements.txt
datasets>=2.14.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
requests>=2.31.0
tqdm>=4.65.0
PyYAML>=6.0
numpy>=1.24.0
python-dotenv>=1.0.0